pyproject.toml
[project]
name = "mmarchand-nsi-github-io"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logic-superfence>=1.2.2",
    "mkdocs>=1.6.1",
    "mkdocs-autorefs>=1.4.3",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-get-deps>=0.2.0",
    "mkdocs-git-revision-date-localized-plugin>=1.5.1",
    "mkdocs-graphviz>=1.5.3",
    "mkdocs-material>=9.7.1",
    "mkdocs-material-extensions>=1.3.1",
    "mkdocs-pyscript>=0.2.1",
    "mkdocs-sqlite-console>=2.1.2",
    "pyyaml>=6.0.3",
]
